# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luka AI - Finnish Master API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SearchRequest {
  word: string;
  apiKey?: string;
  model?: string;
}

export interface OcrRequest {
  imageBase64: string;
  mimeType: string;
  apiKey?: string;
  model?: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  model: "model",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  text: string;
}

export interface ChatRequest {
  message: string;
  imageBase64?: string;
  imageMimeType?: string;
  audioBase64?: string;
  audioMimeType?: string;
  history?: ChatMessage[];
  apiKey?: string;
  model?: string;
}

export interface LukaResponse {
  text: string;
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}

export type HistoryItemType =
  (typeof HistoryItemType)[keyof typeof HistoryItemType];

export const HistoryItemType = {
  search: "search",
  ocr: "ocr",
  chat: "chat",
} as const;

export interface HistoryItem {
  id: string;
  type: HistoryItemType;
  query: string;
  response: string;
  timestamp: string;
}

export interface HistoryResponse {
  items: HistoryItem[];
}

export interface ExportRequest {
  text: string;
  filename?: string;
}