# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luka AI - Finnish Master API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from "./chatMessage";

export interface ChatRequest {
  message: string;
  imageBase64?: string;
  imageMimeType?: string;
  voiceTranscript?: string;
  history?: ChatMessage[];
  apiKey?: string;
  model?: string;
}