# lib/api-zod/src/generated/types/historyItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luka AI - Finnish Master API
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryItemType } from "./historyItemType";

export interface HistoryItem {
  id: string;
  type: HistoryItemType;
  query: string;
  response: string;
  timestamp: Date;
}