# lib/api-zod/src/generated/types/ocrRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luka AI - Finnish Master API
 * OpenAPI spec version: 0.1.0
 */

export interface OcrRequest {
  imageBase64: string;
  mimeType: string;
  apiKey?: string;
  model?: string;
}