# lib/api-zod/src/generated/types/searchRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luka AI - Finnish Master API
 * OpenAPI spec version: 0.1.0
 */

export interface SearchRequest {
  word: string;
  apiKey?: string;
  model?: string;
}